add user middleware
