+++ package.json
@@ -60,7 +60,7 @@
     "chai": "^4.1.2",
     "chai-http": "^3.0.0",
     "coveralls": "^3.0.0",
-    "dictum.js": "^1.0.0",
+    "dictum.js": "^1.0.4",
     "dotenv": "^5.0.0",
     "eslint": "^4.8.0",
     "eslint-config-airbnb-base": "^12.0.2",
